+++ src/popups/StopPlacePopup/StopPlacePopup.js
@@ -74,6 +74,7 @@ const useStopPlaceData = (uic, cartaroUrl) => {
 const formatYesNoData = (data) => {
   if (data === "YES") return "Ja";
   if (data === "NO") return "Nein";
+  if (data === "PARTIALLY") return "Teilweise";
   return data;
 };
 
@@ -88,7 +89,7 @@ const getAccessibility = (value, language, t) => {
     ? note?.split(hardcodedStringForNote)[1]
     : note;
   return (
-    <fieldset key="accessibility">
+    <fieldset key="accessibility" data-testid="stopplace-accessibility">
       <legend>{t("accessibility")}</legend>
       <Typography>{t(formatYesNoData(value?.state))}</Typography>
       {hasTranslatedString && (
@@ -121,12 +122,15 @@ const getAlternativeTransport = (value, language, t) => {
   if (!value) return null;
   return (
     !/^NO|UNKNOWN$/.test(value?.state) && (
-      <fieldset key="alternativeTransport">
+      <fieldset
+        key="alternativeTransport"
+        data-testid="stopplace-alternative-transport"
+      >
         <legend>{t("alternativeTransport")}</legend>
-        <div>{t("Shuttle-Fahrdienst")}</div>
+        <Typography>{t(formatYesNoData(value?.state))}</Typography>
         {typeof value.note === "object"
           ? value.note?.[language] || value.note?.de
-          : value}
+          : value.note}
       </fieldset>
     )
   );
@@ -139,10 +143,17 @@ const getPassengetInformation = (value, t) => {
   });
   return (
     !!entries.length && (
-      <fieldset key="passengerInformation">
+      <fieldset
+        key="passengerInformation"
+        data-testid="stopplace-passengerinfo"
+      >
         <legend>{t("passengerInformation")}</legend>
         {entries.map(([k]) => {
-          return <div key={k}>{t(k)}</div>;
+          return (
+            <div key={k} data-testid={`stopplace-passengerinfo-${k}`}>
+              {t(k)}
+            </div>
+          );
         })}
       </fieldset>
     )
@@ -152,14 +163,14 @@ const getPassengetInformation = (value, t) => {
 const getNote = (value, language, t) => {
   if (!value) return null;
   return (
-    <fieldset key="note">
+    <fieldset key="note" data-testid="stopplace-note">
       <legend>{t("Hinweise zur Haltestelle")}</legend>
       {typeof value === "object" ? value[language] || value.de : value}
     </fieldset>
   );
 };
 
-const getUrl = (value, language, t) => {
+const getUrl = (value, t) => {
   if (!value) return null;
   const url = /^http(s)?:\/\//.test(value) ? value : `https://${value}`;
   let niceVal = value;
@@ -171,7 +182,7 @@ const getUrl = (value, language, t) => {
     niceVal = value;
   }
   return (
-    <fieldset key="url">
+    <fieldset key="url" data-testid="stopplace-url">
       <legend>{t("Weitere Informationen")}</legend>
       <Link href={url}>{niceVal}</Link>
     </fieldset>
@@ -204,7 +215,7 @@ function StopPlacePopup({ feature }) {
       t,
     );
     const note = getNote(infos.note, i18n.language, t);
-    const url = getUrl(infos.url, i18n.language, t);
+    const url = getUrl(infos.url, t);
     const content = [
       accessibility,
       alternativeTransport,
@@ -212,7 +223,7 @@ function StopPlacePopup({ feature }) {
       note,
       url,
     ].filter(Boolean);
-    return content?.length ? content : t(`Keine Daten für diese Station`);
+    return content?.length ? content : t("Keine Daten für diese Station");
   }, [data?.prmInformation, i18n.language, t]);
 
   if (loading) {