+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.26.5](https://github.com/geops/trafimage-maps/compare/v1.26.4...v1.26.5) (2024-01-29)
+
+### Bug Fixes
+
+- **handicap:** prevent popup crash ([#1140](https://github.com/geops/trafimage-maps/issues/1140)) ([8c775fa](https://github.com/geops/trafimage-maps/commit/8c775fa644f00ab540d0c2e5ec107760e44cf631))
+
+### [1.26.4](https://github.com/geops/trafimage-maps/compare/v1.26.3...v1.26.4) (2024-01-26)
+
+### Bug Fixes
+
+- change angular example ([#1136](https://github.com/geops/trafimage-maps/issues/1136)) ([27376e8](https://github.com/geops/trafimage-maps/commit/27376e88c49cf9374eec3831a04fced97d06f0fe))
+
 ### [1.26.3](https://github.com/geops/trafimage-maps/compare/v1.26.2...v1.26.3) (2024-01-25)
 
 ### Bug Fixes